fix: AuctionResultsByFollowedArtists resolver
