+++ src/schema/v2/me/__tests__/auctionResultsByFollowedArtists.test.ts
