+++ src/schema/v2/me/index.ts
@@ -31,7 +31,7 @@ import { SaleArtworksConnectionField } from "../sale_artworks"
 import { ArtistRecommendations } from "./artistRecommendations"
 import { ArtworkRecommendations } from "./artworkRecommendations"
 import ArtworkInquiries from "./artwork_inquiries"
-import AuctionResultsByFollowedArtists from "./auction_results_by_followed_artists"
+import AuctionResultsByFollowedArtists from "./auctionResultsByFollowedArtists"
 import { authentications } from "./authentications"
 import { BankAccounts } from "./bank_accounts"
 import Bidders from "./bidders"
